tests: Extract e2e test utils
Consider creating generic testutils for both e2e and integration tests.
